+++ src/common/INet/Base64.cpp
@@ -124,7 +124,7 @@ int ConvertBase64ToBinary(uchar *Binary, int OutBuf, char *Base64, int InBuf)
 		}
 	}
 
-	return (int)Binary - (int)Start;
+	return Binary - Start;
 }
 
 int ConvertBinaryToBase64(char *Base64, int OutBuf, uchar *Binary, int InBuf)
@@ -190,6 +190,6 @@ int ConvertBinaryToBase64(char *Base64, int OutBuf, uchar *Binary, int InBuf)
 		}
 	}
 
-	return (int)Base64 - (int)Start;
+	return Base64 - Start;
 }
